requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==1.6.0
boto3==1.23.9
botocore==1.26.10
certifi==2022.6.15
cffi==1.15.1
channels==3.0.1
channels-redis==2.4.2
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==3.0.2
Django==3.2.16
django-autoslug==1.9.8
django-cities-light==3.9.2
django-cors-headers==3.7.0
django-environ==0.8.1
django-model-utils==4.2.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-redis==5.2.0
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==13.11.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==5.1.0
incremental==22.10.0
jmespath==0.10.0
msgpack==0.6.2
packaging==21.3
phonenumbers==8.12.48
Pillow==8.4.0
progressbar2==4.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2021.3
redis==3.5.3
requests==2.28.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
toml==0.10.2
tqdm==4.64.0
Twisted==22.10.0
txaio==22.2.1
typing-extensions==4.0.1
Unidecode==1.3.4
urllib3==1.26.9
websocket-client==0.57.0
whitenoise==5.3.0
zipp==3.11.0
zope.interface==5.5.2
